Created helper class with reformat functions
